fix compile error

+++ db/compaction/compaction_job.cc
@@ -840,7 +840,7 @@ Status CompactionJob::Install(const MutableCFOptions& mutable_cf_options) {
   double wal_mb = 0;
   std::vector<std::vector<double>> level_mb;
   for (int i = 0; i < 3; i++) {
-    level_mb.push_back(std::vector<float>(7));
+    level_mb.push_back(std::vector<double>(7));
   }
   for (auto _cfd : *versions_->GetColumnFamilySet()) {
     if (_cfd->IsDropped() || !_cfd->initialized() || _cfd->NumberLevels() < 1) {
